fastapi-cource/pyproject.toml
[project]
name = "fastapi-cource"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
